+++ R/tab_cross.R
@@ -1,9 +1,7 @@
 #' @rdname tab_subset
 #' @export
-tab_cross = function(vr, vrby
-                     , max_levels = getOption("surveytable.max_levels")
-                     , csv = getOption("surveytable.csv")
-) {
+tab_cross = function(vr, vrby, max_levels = getOption("surveytable.max_levels")
+                     ) {
   design = .load_survey()
 
   # Ensure unique name
@@ -18,8 +16,7 @@ tab_cross = function(vr, vrby
     , .getvarname(design, vrby), ")")
   ret = .tab_factor(design = design, vr = newvr
                     , drop_na = FALSE
-                , max_levels = max_levels
-                , csv = csv)
+                    , max_levels = max_levels)
 
   design$variables[,newvr] = NULL
   env$survey = design