+++ R/total_rate.R
@@ -2,7 +2,6 @@
 #'
 #' @param pop population
 #' @param per calculate rate per this many items in the population
-#' @param csv     name of a CSV file
 #'
 #' @return A table
 #' @family tables
@@ -12,8 +11,7 @@
 #' set_survey(namcs2019sv)
 #' total_rate(uspop2019$total)
 total_rate = function(pop
-    , per = getOption("surveytable.rate_per")
-    , csv = getOption("surveytable.csv") ) {
+    , per = getOption("surveytable.rate_per")) {
   assert_that(pop > 0, per >= 1)
   if ( !(per %in% 10^c(2:5)) ) {
     warning("Value of per is not typical: ", per)
@@ -45,5 +43,5 @@ total_rate = function(pop
   attr(m1, "num") = 1:5
   attr(m1, "footer") = attr(mp, "footer")
 
-  .write_out(m1, csv = csv)
+  .finalize_tab(m1)
 }