+++ R/z_test_numeric.R
@@ -1,4 +1,4 @@
-.test_numeric = function(design, vr, vrby, lvl0, alpha, p_adjust, csv, test_title) {
+.test_numeric = function(design, vr, vrby, lvl0, alpha, p_adjust, test_title) {
   assert_that(alpha > 0, alpha < 0.5
               , p_adjust %in% c(TRUE, FALSE))
   if ( !(alpha %in% c(0.05, 0.01, 0.001)) ) {
@@ -36,5 +36,5 @@
 
   .test_table(rT = rT
               , test_name = test_name, test_title = test_title
-              , alpha = alpha, csv = csv)
+              , alpha = alpha)
 }