+++ R/zzz.R
@@ -23,7 +23,6 @@ env = new.env()
     , surveytable.raw = FALSE
     , surveytable.drop_na = FALSE
     , surveytable.max_levels = 20
-    , surveytable.csv = ""
     , surveytable.print = ".print_auto"
     , surveytable.file = ""
     , surveytable.file_show = ""